# CLOUDSTACK-8917 : Instance tab takes long time to load with 12K Vms